adding logging decorator
